fix(cli): bound the engine probe and address review nits
