+++ apps/cli/src/commands/db/start/SIDE_EFFECTS.md
@@ -116,9 +116,9 @@ only spawns on fallback.
 No platform (Management API) routes. The already-running probe issues one Docker Engine
 API request over the active context's local unix socket / named pipe:
 
-| Method | Path                                     | Auth | Request body | Response (used fields)                                                                                |
-| ------ | ---------------------------------------- | ---- | ------------ | ----------------------------------------------------------------------------------------------------- |
-| GET    | `/containers/supabase_db_<project>/json` | —    | —            | status only (Engine-identified 200 → running, 404 → absent; anything else → the spawned-CLI fallback) |
+| Method | Path                                     | Auth | Request body | Response (used fields)                                                                                                                                                          |
+| ------ | ---------------------------------------- | ---- | ------------ | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| GET    | `/containers/supabase_db_<project>/json` | —    | —            | `Api-Version`/`Server` identity headers + status; a 200 body must parse as a JSON object (identity-gated 200 → running, 404 → absent; anything else → the spawned-CLI fallback) |
 
 ## Environment Variables
 
@@ -139,7 +139,7 @@ API request over the active context's local unix socket / named pipe:
 | `SUPABASE_AUTH_EXTERNAL_URL` / `SUPABASE_AUTH_SITE_URL`                                                              | auth migrate job env overrides                                                                                                                                                                                                                                                                                          | no        |
 | `SUPABASE_AUTH_JWT_EXPIRY`                                                                                           | Postgres's `JWT_EXP` env / signing                                                                                                                                                                                                                                                                                      | no        |
 | `SUPABASE_EXPERIMENTAL` (or `--experimental`)                                                                        | fresh volume + no pg-delta: applies `db.migrations.schema_paths` files instead of `migrations/*.sql`                                                                                                                                                                                                                    | no        |
-| `DOCKER_HOST` / `DOCKER_CONTEXT` / `DOCKER_TLS_VERIFY` / `DOCKER_CERT_PATH` / `DOCKER_API_VERSION` / `DOCKER_CONFIG` | Read (ambient shell OR a project `.env`/`.env.<env>`/`.env.local` file, installed into the process environment before any Docker work) to pick the Docker daemon this whole command talks to                                                                                                                            | no        |
+| `DOCKER_HOST` / `DOCKER_CONTEXT` / `DOCKER_TLS_VERIFY` / `DOCKER_CERT_PATH` / `DOCKER_API_VERSION` / `DOCKER_CONFIG` | Read from the ambient shell environment to pick the Docker daemon this whole command talks to (project dotenv files deliberately never override Docker client keys)                                                                                                                                                     | no        |
 | `SUPABASE_USE_SLIM_IMAGES`                                                                                           | resolves the current Dockerfile pin (and majors 13/15's published slim PG15 pin, `15.14.1.167`) and PG15+ realtime/storage/auth migrate-job images from the slim `ghcr.io/supabase/cli` builds (`true`/`1` enable); historical `.temp` pins, PG14, OrioleDB, and flag-off majors 13/15 (`15.8.1.085`) stay on docker.io | no        |
 
 `--network-id` (a global CLI flag, not an environment variable — `command-internal/global-flags.ts`)
